pyproject.toml
[tool.poetry]
name = "arxiv-tex2pdf"
version = "0.1.0"
description = "Compiling TeX to PDF as a service"
authors = ["arxiv.org"]
packages = [
    { include = "tex2pdf" }
]
include = ["LICENSE", "README.md"]

[tool.poetry.dependencies]
python = "^3.11"
chardet = "^5.0"
pydantic = "==1.10.*"
ruamel-yaml = "^0.18.5"
toml = "^0.10.2"
tomli_w = "^1.0"

[tool.poetry.group.dev.dependencies]
ruff = "*"
pytest = "^8.3.3"
mypy = "*"
mypy-extensions = "*"
types-toml = "^0.10.8.20240310"

[tool.ruff]
line-length = 120
extend-exclude = [
    "tex2pdf/preflight_parser/tests",
    "tex2pdf/zerozeroreadme/tests",
    "tex2pdf/tex_inspection/tests",
]

lint.select = [
    "F", # Pyflakes
    "D", # pydocstyle
    "E", # pycodestyle errors
    # "W", # pycodestyle warnings
    # "C90", # McCabe
    "I",   # isort
    "UP",  # pyupgrade
    "PLC",
    "PLE",
    "PLW", # Pylint
    # "PLR", # Pylint refactoring
    "PD",  # pandas checks
    "NPY", # numpy checks
    "RUF", # ruff internal check
]
lint.ignore = [
    "D105", # no docstring in magic method like __str__
    "D107", # no docstring in __init__
]


output-format = "grouped"

# Assume Python 3.11
# Note: helps prevent breaking autofixes from, e.g., pyupgrade
target-version = "py311"

[tool.ruff.lint.pydocstyle]
convention = "pep257"

[tool.black]
line-length = 120

[tool.pylint]
max-line-length = 120

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


